raw operf support added

+++ oprofile-init.sh
@@ -10,22 +10,40 @@ if [ $# -ne 1 ]; then
 fi
 test_dir="$1"
 
-if [ "$(which opcontrol 2>/dev/null)" != "" ]; then
-    echo "starting oprofile"
+#read config file
+use_operf="$( cat "$test_dir/wrstat.config" | grep "oprofile_use_operf" | cut -d " "  -f 2-)"
+use_operf=${use_operf,,}
+vmlinux="$( cat "$test_dir/wrstat.config" | grep "oprofile_vmlinux" | cut -d " "  -f 2-)"
 
-    rm -rf "$test_dir/oprofile_data/" &>/dev/null
-    opcontrol --reset
-    opcontrol --deinit
-    modprobe oprofile timer=1
-    echo 0 > /proc/sys/kernel/nmi_watchdog
-    opcontrol --separate=cpu #--separate=none
+#check if we should use operf or opcontrol (deprecated)
+if [ $use_operf == "true" ]; then
+    if [ "$(which operf 2>/dev/null)" != "" ]; then
+        #starting operf
+        if [ "$vmlinux" == "" ];then
+            operf --systemwide --separate-cpu --session-dir="$test_dir/oprofile_data/"&
+        else
+            ln -sf "$vmlinux" "$test_dir/vmlinux"
+            operf --systemwide --separate-cpu --vmlinux="$test_dir/vmlinux" --session-dir="$test_dir/oprofile_data/"&
+        fi
+        echo $! > "$test_dir/operf.pid"
+    fi
+#opcontrol
+else
+    if [ "$(which opcontrol 2>/dev/null)" != "" ]; then
+        echo "starting oprofile"
 
-    vmlinux="$( cat "$test_dir/wrstat.config" | grep "oprofile_vmlinux" | cut -d " "  -f 2-)"
-    if [ "$vmlinux" == "" ]
-    then
-        opcontrol --start --no-vmlinux --session-dir="$test_dir/oprofile_data/"
-    else
-        ln -sf "$vmlinux" "$test_dir/vmlinux"
-        opcontrol --start --vmlinux="$test_dir/vmlinux" --session-dir="$test_dir/oprofile_data/"
+        rm -rf "$test_dir/oprofile_data/" &>/dev/null
+        opcontrol --reset
+        opcontrol --deinit
+        modprobe oprofile timer=1
+        echo 0 > /proc/sys/kernel/nmi_watchdog
+        opcontrol --separate=cpu #--separate=none
+
+        if [ "$vmlinux" == "" ];then
+            opcontrol --start --no-vmlinux --session-dir="$test_dir/oprofile_data/"
+        else
+            ln -sf "$vmlinux" "$test_dir/vmlinux"
+            opcontrol --start --vmlinux="$test_dir/vmlinux" --session-dir="$test_dir/oprofile_data/"
+        fi
     fi
 fi
